test using python 3.12
STT-1517
