+++ .github/workflows/tests.yml
@@ -10,7 +10,7 @@ jobs:
   server:
     strategy:
       matrix:
-        python-version: ['3.10']
+        python-version: ['3.10', '3.12']
     uses: superdesk/superdesk/.github/workflows/server.yml@develop
     with:
       python-version: ${{ matrix.python-version }}
@@ -18,7 +18,7 @@ jobs:
   pytest:
     strategy:
       matrix:
-        python-version: ['3.10']
+        python-version: ['3.10', '3.12']
     uses: superdesk/superdesk/.github/workflows/pytest.yml@develop
     with:
       python-version:  ${{ matrix.python-version }}
@@ -26,7 +26,7 @@ jobs:
   behave:
     strategy:
       matrix:
-        python-version: ['3.10']
+        python-version: ['3.10', '3.12']
     uses: superdesk/superdesk/.github/workflows/behave.yml@develop
     with:
       python-version:  ${{ matrix.python-version }}
