+++ server/dev-requirements.txt
@@ -1,6 +1,7 @@
 -r requirements.txt
 
 black==25.11.0
+behave==1.2.6
 mypy==1.19.0
 flake8==7.3.0
 pep8==1.7.1
@@ -12,5 +13,5 @@ responses==0.25.8
 aioresponses==0.7.8
 requests-mock==1.12.1
 
-types-requests
-types-pytz
+types-requests==2.31.0.6
+types-pytz~=2025.2
